Move test with mallory as a token
